Refactors bilinear resize to use fixed-point math
Replaces floating-point calculations in the bilinear image resize implementation with Q15 fixed-point arithmetic for improved performance and consistency, particularly on platforms with limited floating-point support. Adds helper functions for fixed-point multiplication and clamping, and ensures all lookup and interpolation steps use integer math. Also fixes a timestamp calculation to avoid potential overflow by explicitly casting operands.

+++ test/util/util.c
@@ -296,7 +296,7 @@ uint64_t sgl_test_get_timestamp_us(uint64_t start_us)
     uint64_t val;
 
     (void)clock_gettime(CLOCK_MONOTONIC, &timespec);
-    val = (timespec.tv_sec * 1000000ULL) + (timespec.tv_nsec / 1000ULL);
+    val = ((uint64_t)timespec.tv_sec * 1000000ULL) + ((uint64_t)timespec.tv_nsec / 1000ULL);
 
     if (0ULL < start_us) {
         val -= start_us;
